==> Data Analytics/Data Analytics on Multiple channels/1_Extracting_data_from_Youtube_API.ipynb <==
# Aim and Objective

* Getting to know Youtube API and how to obtain video data.
* Analyzing video data and verify different common "myths" about what makes a video do well on Youtube, for example:
* Does the number of likes and comments matter for a video to get more views?
* Does the video duration matter for views and interaction (likes/ comments)?
* To Analyze the video metrics such as views, duration, upload schedule, likes, comments, tags, and upload frequency for popular music channels ('Aditya Music', 'SonyMusicSouthVEVO', 'T-Series', 'Saregama Music', and 'Zee Music Company') to determine their impact on video performance.
* To visualized and compared best/worst and shortest/longest performing videos and examined weekly/yearly upload schedules to identify how these metrics affect views.



# Steps:
1. Data Extraction using Youtube API
2. Data Preprocessing
3. Feature Selection and Adding additional features
4. Exploratory Data Analysis
5. Observations & Conclusions

# Dataset

In regards to music channels, I discovered that there are not many available datasets online that are suitable for this project.

* collected data from the following music channels: 'Aditya Music', 'SonyMusicSouthVEVO', 'T-Series', 'Saregama Music', and 'Zee Music Company'. The parameters that I included in my data collection were 'channelTitle', 'title', 'tags', 'publishedAt', 'viewCount', 'likeCount', 'commentCount', 'duration', 'tagcount', 'publishDayName', 'durationSecs', and 'publishYear'.
* I created my own dataset using the [Google Youtube Data API version 3.0](https://developers.google.com/youtube/v3)

#Code

from googleapiclient.discovery import build
import pandas as pd
from IPython.display import JSON

import seaborn as sns

api_key='My Youtube API key'

channel_id= ['UC_A7K2dXFsTMAciGmnNxy-Q', #Saregama Music
             'UCq-Fj5jknLsUf-MWSy4_brA', #T-series
             'UCTNtRdBAiZtHP9w7JinzfUg', #SonyMusicSouth
             'UCNApqoVYJbYSrni4YsbXzyQ', #Aditya Music
             'UCFFbwnve3yF62-tVXkTyHqg'  #Zee Music
             ] 

api_service_name = "youtube"
api_version = "v3"


# Get credentials and create an API client
youtube = build(
    api_service_name, api_version, developerKey=api_key)

request= youtube.channels().list(
      part="snippet,contentDetails,statistics",
      id=','.join(channel_id)
  )
response= request.execute()

JSON(response)


def get_channel_stats(youtube, channel_id):
  all_data=[]

  request= youtube.channels().list(
      part="snippet,contentDetails,statistics",
      id=','.join(channel_id)
  )
  response= request.execute()

  for item in response['items']:
    data= {'channelName':item['snippet']['title'],
           'subscribers': item['statistics']['subscriberCount'],
           'views':item['statistics']['viewCount'],
           'totalViews': item['statistics']['videoCount'],
           'playlistId': item['contentDetails']['relatedPlaylists']['uploads']
    }

    all_data.append(data)

  return(pd.DataFrame(all_data))

channel_stats= get_channel_stats(youtube, channel_id)

channel_stats

def get_video_ids_for_playlists(youtube, playlist_ids):
    video_ids = []
    for playlist_id in playlist_ids:
        request = youtube.playlistItems().list(
            part="contentDetails",
            playlistId=playlist_id,
            maxResults=50
        )
        response = request.execute()
        for item in response["items"]:
            video_ids.append(item["contentDetails"]["videoId"])
        while "nextPageToken" in response:
            page_token = response["nextPageToken"]
            response = youtube.playlistItems().list(
                part="contentDetails",
                playlistId=playlist_id,
                maxResults=50,
                pageToken=page_token
            ).execute()
            for item in response["items"]:
                video_ids.append(item["contentDetails"]["videoId"])
    return video_ids

playlist_ids = ['UUNApqoVYJbYSrni4YsbXzyQ', 'UUTNtRdBAiZtHP9w7JinzfUg', 'UUq-Fj5jknLsUf-MWSy4_brA', 'UU_A7K2dXFsTMAciGmnNxy-Q', 'UUFFbwnve3yF62-tVXkTyHqg']

# Call the function to get the video IDs for all videos in the playlists
video_ids = get_video_ids_for_playlists(youtube, playlist_ids)


len(video_ids)

def get_video_details(youtube, video_ids):
        
    all_video_info = []
    
    for i in range(0, len(video_ids), 50):
        request = youtube.videos().list(
            part="snippet,contentDetails,statistics",
            id=','.join(video_ids[i:i+50])
        )
        response = request.execute() 

        for video in response['items']:
            stats_to_keep = {'snippet': ['channelTitle', 'title', 'description', 'tags', 'publishedAt'],
                             'statistics': ['viewCount', 'likeCount', 'favouriteCount', 'commentCount'],
                             'contentDetails': ['duration', 'definition', 'caption']
                            }
            video_info = {}
            video_info['video_id'] = video['id']

            for k in stats_to_keep.keys():
                for v in stats_to_keep[k]:
                    try:
                        video_info[v] = video[k][v]
                    except:
                        video_info[v] = None

            all_video_info.append(video_info)

    return pd.DataFrame(all_video_info)

video_df= get_video_details(youtube, video_ids)
len(video_df)

video_df.head()

video_df.isnull().sum()

video_df.columns

final_df= video_df.drop(['video_id','description','caption','favouriteCount','definition'],axis=1)

final_df.to_csv('final_df.csv', index=False, header=True)

final_df.shape

final_df.isnull().sum()

final_df[final_df['title'].isnull()]

final_df.tail()
