==> Data Analytics/Data Analytics on Multiple channels/2_Data_preprocessing.ipynb <==
# **Preprocessing** involves removing duplicates, converting necessary values to numerical from object datatype, finding Null values and dealing with them.

import pandas as pd

video_df= pd.read_csv('/content/final_df.csv')

# Drop duplicate rows
video_df = video_df.drop_duplicates()

video_df.head()

video_df.shape

video_df[video_df['title'].isnull()]

video_df['channelTitle'].unique()

video_df = video_df[video_df['channelTitle'].map(video_df['channelTitle'].value_counts()) > 1]

video_df['channelTitle'].unique()

video_df.reset_index(inplace = True, drop = True)

video_df.tail()

len(video_df)

video_df.isnull().sum()

video_df.dtypes

#Changing 'viewCount', 'likeCount','commentCount' into Numeric Format
numeric_cols= ['commentCount']
video_df[numeric_cols]=video_df[numeric_cols].apply(pd.to_numeric,errors='coerce',axis=1)

video_df.dtypes

video_df['likeCount'].isnull().sum(), video_df['commentCount'].isnull().sum()

# select rows where both likecount and commentcount are NULL
null_rows = video_df[(video_df['likeCount'].isnull()) & (video_df['commentCount'].isnull())]

# print the selected rows
pd.DataFrame(null_rows)

len(video_df),len(null_rows)

#Deleting this 15 rows as they can't impact much on dataset
video_df= video_df.drop(null_rows.index)
len(video_df)

video_df.head()

cmnt_null=video_df[video_df['commentCount'].isnull()]
video_df=video_df.drop(cmnt_null.index)
len(video_df)

video_df['likeCount'].describe()

video_df.head()

video_df.isnull().sum()

mean_like= video_df['likeCount'].mean()
mean_like

video_df['likeCount'].fillna(mean_like,inplace=True)

video_df.dtypes

video_df.to_csv('video_df.csv', index=False, header=True)
