==> Data Analytics/Data Analytics on Multiple channels/3_Feature_Selection.ipynb <==
# **Feature Selection** involves adding tag count, duration in Seconds, published weekday and year of a video.

import pandas as pd

# Mount Google Drive
from google.colab import drive
drive.mount('/content/drive',force_remount=True)

# Define file path of CSV file in Google Drive
file_path = "/content/drive/MyDrive/video_df.csv"

# Load CSV file into a pandas dataframe
df = pd.read_csv(file_path)

# Print first few rows of the dataframe
df.head()

df[['viewCount', 'likeCount', 'commentCount']] = df[['viewCount', 'likeCount', 'commentCount']].astype(int)

df.dtypes

df.head()

df['channelTitle'].unique()

df['tags'].isnull().sum()

df['tags'][123]

max(df.tagcount)

#Adding Tag count
df['tagcount'] = df['tags'].apply(lambda x: 0 if pd.isna(x) else len(x.split("|")))

!pip install isodate

#Duration into Duration in seconds
from isodate import parse_duration

df['durationSecs']= df['duration'].apply(lambda x: parse_duration(x))
df['durationSecs']=df['durationSecs'].astype('timedelta64[s]')

df[['durationSecs','duration']]

#published Year
from dateutil import parser
# Extract year from publishedAt column and add to DataFrame
df['publishYear'] = df['publishedAt'].apply(lambda x: x.strftime("%Y"))

df['publishDayName'] = df['publishedAt'].apply(lambda x: x.strftime("%A"))

df.head()

df.columns

final_df= df.drop(['tags','publishedAt','duration'],axis=1)

final_df.head()

final_df.shape

final_df.to_csv('Dataset_for_Analysis.csv',index=False, header= True)
